Bug fixes, debugging helpers, and a new Exception handler.
Also, although Schema 3.0.9.5 was "finalized" in 992aef9, there were a
couple things I forgot. I'm not bothering to update the schema version
number again, because 3.0.9.5 has not hit the outside world yet (and by
that I mean the lappy).

So consider *this* submit the real 3.0.9.5 schema.

-----------------------------------------------------------------------
TESTING
-----------------------------------------------------------------------

- Only real visible thing is: creating a new database should now create
  all the tables.

-----------------------------------------------------------------------
DETAILS
-----------------------------------------------------------------------

- Classes/Event.cs (modified)
  * FK constraint enforcement fixes...
  * Made ReminderId and ScheduleId nullable.

- Classes/Schedule.cs (modified)
  * FK constraint enforcement fixes...
  * Made ScheduleId nullable.

- Classes/ScheduledEventCollection.cs (modified)
  * Fixed bug where this would die if Events table were empty.

- Classes/Timekeeper.cs (modified)
  * Updated Exception handler to print out more information (including
    the all-important line number) when normal exceptions are called.
  * Completely replaced the stack-trace code. Instead of using the
    built-in Environment.StackTrace, wrote my own to give me more
    control over the output (primarily, stopping the stack trace as
    soon as I don't have access to the source code anymore).
  * Updated StringToDate(string) to return DateTimeOffset.MinValue if
    the string argument was null.

- Forms/Main.Browser.cs (modified)
  * Added debugging code to help me zero in on Issue #1404.

- Forms/Main.Designer.cs (modified)
  * Renamed a couple items on the Tools menu.

- Forms/Main.resx (modified)
  * Updated Tray icon (finally).

- Forms/Shared/Reminder.cs (modified)
- Forms/Shared/Schedule.cs (modified)
  * Made ReminderId and ScheduleId nullable.

- Forms/Tools/Countdown.Designer.cs (modified)
  * Began stubbing in Todo menu.

- Forms/Tools/Countdown.cs (modified)
  * Event window floats independent of this window.
  * Began stubbing in Todo menu.

- Forms/Tools/EventDetail.cs (modified)
  * Made ReminderId and ScheduleId nullable.

- Properties/Resources.Designer.cs (modified)
- Properties/Resources.resx (modified)
- Timekeeper3.csproj (modified)
  * Added CalendarView and PunchCardView to the resources.

- Resources/SQL/3.0/Event.Create.sql (modified)
  * Made ReminderId and ScheduleId nullable.

-----------------------------------------------------------------------
WORKFLOW
-----------------------------------------------------------------------

Project Name....: Timekeeper
Components......: Multiple
Target Version..: 3.0
Type of Change..: Bug Fix

Tickets Addressed:
  #1383: 3.0.511.1: Event exception.
  #1402: 3.0.513.1: Countdown should also count down to Todo Item due dates.
  #1404: 3.0.513.1: Data loss.

Tickets Closed:
  #1403: Timekeeper exception handler should report a line number.
  #1407: 3.0.513.1: New databases missing new tables.

+++ timekeeper/Forms/Tools/EventDetail.cs
@@ -131,14 +131,14 @@ namespace Timekeeper.Forms.Tools
 
                     // CAREFUL: This may whomp things in a way we do not want them whomped
                     CurrentEvent.ReminderId = 0;
-                    CurrentEvent.Reminder = new Classes.Reminder(CurrentEvent.ReminderId);
+                    CurrentEvent.Reminder = new Classes.Reminder((long)CurrentEvent.ReminderId);
                     //Common.Info("User don't want no reminder"); 
                     break;
 
                 case DialogResult.OK:
                     // CAREFUL: This may whomp things in a way we do not want them whomped
                     CurrentEvent.ReminderId = ReminderForm.CurrentReminder.ReminderId;
-                    CurrentEvent.Reminder = new Classes.Reminder(CurrentEvent.ReminderId);
+                    CurrentEvent.Reminder = new Classes.Reminder((long)CurrentEvent.ReminderId);
                     //Common.Info("User want reminder!"); 
                     break;
 
@@ -160,7 +160,7 @@ namespace Timekeeper.Forms.Tools
             switch (ScheduleForm.DialogResult) {
                 case DialogResult.OK:
                     CurrentEvent.ScheduleId = ScheduleForm.CurrentSchedule.ScheduleId;
-                    CurrentEvent.Schedule = new Classes.Schedule(CurrentEvent.ScheduleId);
+                    CurrentEvent.Schedule = new Classes.Schedule((long)CurrentEvent.ScheduleId);
                     CurrentEvent.NextOccurrenceTime = ScheduleForm.ExternalEventNextOccurrence;
                     EventNextOccurrence.Value = CurrentEvent.NextOccurrenceTime.DateTime;
                     break;
